Add ecosystem management section
