+++ packages/web/docs/src/components/CheckIcon.tsx
@@ -1,0 +1,13 @@
+export function CheckIcon(props: React.SVGProps<SVGSVGElement>) {
+  return (
+    <svg width="20" height="20" fill="none" {...props}>
+      <path
+        d="M16.667 5 7.5 14.167 3.333 10"
+        stroke="currentColor"
+        strokeWidth="1.5"
+        strokeLinecap="square"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
